requirements.txt
streamlit
python-dotenv
langchain-community
langchain-chroma
langchain-text-splitters
langchain-huggingface
langchain-groq
pypdf
nltk
faiss-cpu
sentence-transformers